+++ src/types/util.ts
@@ -54,14 +54,3 @@ export type Store = {
   alerts: AlertsMap;
   mnemonics: MnemonicsMap;
 };
-
-export type GenericDataType = Contact | Alert | Mnemonic;
-export type GenericDataMapType = ContactsMap | AlertsMap | MnemonicsMap;
-
-export type GetDataTypeFromKey<T extends keyof Store> = T extends 'contact'
-  ? Contact
-  : T extends 'alert'
-  ? Alert
-  : T extends 'mnemonic'
-  ? Mnemonic
-  : never;
